fix "Enable DSCP Fix", and make it MultiWan aware.
When this code was last moved, it was inside one too many braces and put into an if statement (TTL related)
where it would (practically) never execute. It is now MultiWan aware, and will be applied to every
Wan interface when enabled.

The DSCP fix is supposed to improve wifi speeds and is discussed in this thread.
http://linksysinfo.org/index.php?threads/critical-dscp-bug-affecting-wifi-download-speeds-on-comcast-cisco-e-series-and-others.69006/