~/router/rc/wan.c: dnsmasq process was receiving a second SIGINT signal. Instead of triggering another DNSSEC time checking, it was killing process
function to start dnsmasq service ~/rc/services.c actually does the same thing once dnsmasq is started, and does it better, by verifying if system time is greater than the epoch.
Probably at that time such piece of code was fit in there, SIGHUP did a bunch of stuff so receiving it multiple times wasn't really a problem, but receiving SIGINT twice from 2.79 and onwards is harmful.

Thanks to koitsu for the very helpful insight.