Makefile: disable RAID (mdadm binary)
While having RAID-0/1 feature is nice on paper, the USB controller would be a bottleneck, as it is not suitable for this task, specially on a consumer router.
In five years watching the forums, never seen anyone actually using it, and mdadm functions are not exposed on GUI as this is meant for advanced users.
Also these advanced users wouldn't do RAID on a consumer router, for numerous reasonsm, so I believe we'd better sunset this feature.

Everything else for now, remains on the tree. If someone wants mdadm, just compile with RAID=y for AIO target (z).