bridge: Fix IPv6 multicast snooping by correcting offset in MLDv2
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/net/bridge/br_multicast.c?h=v2.6.38&id=649e984d00416cb1a254fdbebd6d3f9fa01c32fa

We actually want a pointer to the grec_nsrcr and not the following
field. Otherwise we can get very high values for *nsrcs as the first two
bytes of the IPv6 multicast address are being used instead, leading to
a failing pskb_may_pull() which results in MLDv2 reports not being parsed.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>
Signed-off-by: David S. Miller <davem@davemloft.net>