bridge: mcast snooping, fix length check of snooped MLDv1/2
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?h=v2.6.39&id=ff9a57a62afbbe2d0f3a09af321f1fd7645f38a5

"len = ntohs(ip6h->payload_len)" does not include the length of the ipv6 header itself,
which the rest of this function assumes, though. This leads to a length check less
restrictive as it should be in the following line for one thing. For another, it very likely
leads to an integer underrun when substracting the offset and therefore
to a very high new value of 'len' due to its unsignedness.
This will ultimately lead to the pskb_trim_rcsum() practically never being called,
even in the cases where it should.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>
Signed-off-by: David S. Miller <davem@davemloft.net>