rc: network.c - change blink behaviour / start
If additional virtual wireless inferfaces are defined, do not use blink and only turn the WLAN LEDs on or off (summary of each wlan unit)

And also a Bug-Fix for: If a virtual wireless Interface was defined and disabled, WLAN LEDs blinking did not work at all! stealth mode... ;-)

thx at edrikk for finding this problem.

tested with RT-N15U and R7000, working now