rc: led.c: - extend led cmd to react properly with blink turned on
Fix for: with blink turned on, turning on/off the WLAN LEDs with led cmd (for example: led wlan off)
does not "work" (at least just a few milliseconds until blink turns the LED back on/off).
Now: if the user turns a WLAN LED on OR off, stop/kill blink.