RT-N18U: make sure to enable gpio 13 for WiFi IC after re-/boot (only for safety)
Make sure that GPIO Pin 13 is always on --> it is connected to the WiFi IC (Power supply?);
otherwise signal will be very weak!
So far this is no change, the Asus driver for Tomato RT-N18U turns GPIO 13 on (HIGH) at startup.
But to make sure, add it. (align also to dd wrt)

Additional Reference:
https://svn.dd-wrt.com/changeset/30263/src/router/services/sysinit/sysinit-northstar.c
gpio13 is somehow connected to the wireless chipset. so we need to keep it enabled, otherwise signal will be weak on the Asus RT-N18U

tested with RT-N18U, working (like before)